pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vct"
version = "0.14.0"
description = "Voice Canine Trainer + RoboDog Brain (LLM-first control, reward GPIO, ethics guard)"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Ярослав / DexterX" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.2",
  "pyttsx3>=2.90",
  "gTTS>=2.5.1",
  "playsound>=1.3.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0.2",
  "paho-mqtt>=2.1.0",
  "prometheus-client>=0.20.0",
  "openai-whisper>=20231117",
]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["vct"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
strict_optional = true
plugins = []
exclude = ["tests/fixtures"]

[[tool.mypy.overrides]]
module = [
  "fastapi",
  "fastapi.*",
  "pydantic",
  "pydantic.*",
  "yaml",
  "yaml.*",
  "gtts",
  "playsound",
  "pyttsx3",
  "whisper",
  "paho",
  "paho.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --cov=vct --cov-report=term-missing --cov-report=xml"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["vct"]

[tool.coverage.report]
fail_under = 80
show_missing = true
